# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BugShot API Documentation",
        "description": "OpenApi documentation using swagger for BugShot project.",
        "contact": {
            "email": "info@it-michel.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.test/api/v1"
        }
    ],
    "paths": {
...
        "/projects/{project_id}/statuses": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "All statuses.",
                "operationId": "allStatuses",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Project/properties/id"
                        }
                    },
                    {
                        "name": "include-bugs",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-screenshots",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-comments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-bug-users",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachment-base64",
                        "in": "header",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Status"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Status"
                ],
                "summary": "Store one status.",
                "operationId": "storeStatus",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Project/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation",
                                    "order_number"
                                ],
                                "properties": {
                                    "designation": {
                                        "description": "The message",
                                        "type": "string"
                                    },
                                    "order_number": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{project_id}/statuses/{status_id}": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Show one status.",
                "operationId": "showStatus",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Project/properties/id"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Status/properties/id"
                        }
                    },
                    {
                        "name": "include-bugs",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-screenshots",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-comments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-bug-users",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachment-base64",
                        "in": "header",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Status"
                ],
                "summary": "Update a status.",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Project/properties/id"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Status/properties/id"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation",
                                    "order_number"
                                ],
                                "properties": {
                                    "designation": {
                                        "description": "The message",
                                        "type": "string"
                                    },
                                    "order_number": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Status"
                ],
                "summary": "Delete a status.",
                "operationId": "deleteStatus",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Project/properties/id"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Status/properties/id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Status": {
                "properties": {
                    "id": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "designation": {
                        "description": "The status name.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "order_number": {
                        "description": "The status name.",
                        "type": "integer",
                        "format": "int32"
                    },
                    "project_id": {
                        "description": "The id of the project to which the object belongs.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "created_at": {
                        "description": "The creation date.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The last date when the resource was changed.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "The deletion date.",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}